Abstracted display of landing page outside of the app, to test connected component witout mocking redux.
